[XGPU] Fix value casting

+++ Xenon/Core/XGPU/CommandProcessor.cpp
@@ -25,7 +25,7 @@ void CommandProcessor::CPUpdateRBSize(size_t newSize) {
   if ((newSize & CP_RB_CNTL_RB_BUFSZ_MASK) == 0)
     return;
 
-  cpRingBufferSize = static_cast<size_t>(uint32_t(1) << (newSize + 3));
+  cpRingBufferSize = static_cast<size_t>(1u << (newSize + 3));
   LOG_DEBUG(Xenos, "CP: Updating RingBuffer Size: {:#x}", cpRingBufferSize.load());
 }
